+++ contracts/storage/CollateralManagerStorage.sol
@@ -25,3 +25,10 @@ abstract contract CollateralManagerStorageV1 {
 
     uint256 internal _collateralValueDust;
 }
+
+abstract contract CollateralManagerStorageV2 is CollateralManagerStorageV1 {
+    // key: trader address, value: whitelisted debt threshold
+    mapping(address => uint256) internal _whitelistedDebtThresholdMap;
+
+    uint256 internal _totalWhitelistedDebtThreshold;
+}